refactor: modularize game logic
